Move clear_poetry_cache fixture to conftest.py
