+++ tests/test_pip_repositories.py
@@ -1,7 +1,5 @@
 import base64
 import os
-import shutil
-import sys
 import tarfile
 
 from io import BytesIO
@@ -13,9 +11,6 @@ import pytest
 import requests
 import requests_mock
 
-from filelock import FileLock
-
-from conda_lock._vendor.poetry.locations import DEFAULT_CACHE_DIR
 from conda_lock.conda_lock import DEFAULT_LOCKFILE_NAME, run_lock
 from conda_lock.lockfile import parse_conda_lock_file
 from tests.test_conda_lock import clone_test_dir
@@ -135,25 +130,6 @@ def configure_auth(monkeypatch):
     monkeypatch.setenv("PIP_PASSWORD", _PRIVATE_REPO_PASSWORD)
 
 
-@pytest.fixture()
-def cleared_poetry_cache(tmp_path_factory, testrun_uid: str):
-    """Ensure no concurrency for tests that rely on the cache being cleared"""
-    # testrun_uid comes from xdist <https://stackoverflow.com/a/62765653>
-    # The idea for using FileLock with the base temp directory comes from
-    # <https://pytest-xdist.readthedocs.io/en/latest/how-to.html#making-session-scoped-fixtures-execute-only-once>
-    root_tmp_dir = tmp_path_factory.getbasetemp().parent
-    testrun_lockfile = root_tmp_dir / f".conda_lock_pytest_{testrun_uid}.lock"
-    with FileLock(testrun_lockfile):
-        # Use `pytest -s` to see these messages
-        print(
-            f"Clearing {DEFAULT_CACHE_DIR} based on lock {testrun_lockfile}",
-            file=sys.stderr,
-        )
-        clear_poetry_cache()
-        yield
-        print(f"Releasing lock {testrun_lockfile}", file=sys.stderr)
-
-
 def test_it_uses_pip_repositories_with_env_var_substitution(
     monkeypatch: "pytest.MonkeyPatch",
     conda_exe: str,
@@ -197,18 +173,3 @@ def test_it_uses_pip_repositories_with_env_var_substitution(
         "Password environment variable was not respected, See full lock-file:\n"
         + lockfile_content
     )
-
-
-def clear_poetry_cache() -> None:
-    # We are going to rmtree the cache directory. Let's be extra careful to make
-    # sure we only delete a directory named "pypoetry-conda-lock" or one of its
-    # subdirectories.
-    to_delete = DEFAULT_CACHE_DIR.resolve()
-    assert to_delete.name == "pypoetry-conda-lock" or (
-        to_delete.parent.name == "pypoetry-conda-lock" and to_delete.name == "Cache"
-    )
-    # Do another independent check that triggers even if we're in optimized mode
-    if "pypoetry-conda-lock" in to_delete.parts:
-        shutil.rmtree(DEFAULT_CACHE_DIR, ignore_errors=True)
-    else:
-        raise RuntimeError(f"Refusing to delete {to_delete} as it does not look right")
